API

NamedTrajectory methods

Base.:== — Method

:(==)(traj1::NamedTrajectory, traj2::NamedTrajectory)

Check if trajectories are equal w.r.t. using Base.:(==)

Base.copy — Method

copy(::NamedTrajectory)

Returns a copy of the trajectory.

Base.getindex — Method

getindex(traj, ts::AbstractVector{Int})::Vector{KnotPoint}

Returns the knot points at times ts.

Base.getindex — Method

getindex(traj, t::Int)::KnotPoint

Returns the knot point at time t.

Base.getindex — Method

getindex(traj, symb::Symbol)

Dispatches indexing of trajectories as either accessing a component or a property via getproperty.

Base.getproperty — Method

getproperty(traj, symb::Symbol)

Returns the component of the trajectory with name symb or the property of the trajectory with name symb.

Base.isequal — Method

isequal(traj1::NamedTrajectory, traj2::NamedTrajectory)

Check if trajectories are equal w.r.t. data using Base.isequal

Base.lastindex — Method

lastindex(traj::NamedTrajectory)

Returns the final time index of the trajectory.

Base.setproperty! — Method

setproperty!(traj, name::Symbol, val::Any)

Dispatches setting properties of trajectories as either setting a component or a property via setfield! or update!.

Base.size — Method

size(traj::NamedTrajectory) = (dim = traj.dim, T = traj.T)

NamedTrajectories.MethodsNamedTrajectory.add_component! — Method

add_component!(traj, name::Symbol, data::AbstractVecOrMat; type={:state, :control})

Add a component to the trajectory.

NamedTrajectories.MethodsNamedTrajectory.components — Method

components(::NamedTrajectory)

Returns a NamedTuple containing the names and corresponding data matrices of the trajectory.

NamedTrajectories.MethodsNamedTrajectory.get_times — Method

get_times(traj)::Vector{Float64}

Returns the times of a trajectory as a vector.

NamedTrajectories.MethodsNamedTrajectory.get_timesteps — Method

get_timesteps(::NamedTrajectory)

Returns the timesteps of a trajectory as a vector.

NamedTrajectories.MethodsNamedTrajectory.remove_component — Method

remove_component(traj, name::Symbol)

Remove a component from the trajectory.

NamedTrajectories.MethodsNamedTrajectory.remove_components — Method

remove_components(traj, names::Vector{Symbol})

Remove a set of components from the trajectory.

NamedTrajectories.MethodsNamedTrajectory.update! — Method

update!(traj, name::Symbol, data::AbstractMatrix{Float64})

Update a component of the trajectory.

NamedTrajectories.MethodsNamedTrajectory.update_bound! — Function

update_bound!(traj, name::Symbol, data::Real)
update_bound!(traj, name::Symbol, data::AbstractVector{<:Real})
update_bound!(traj, name::Symbol, data::Tuple{R, R} where R <: Real)

Update the bound of a component of the trajectory.

random trajectory methods

Base.rand — Method

rand(
    ::Type{NamedTrajectory},
    T::Int;
    timestep_value::Float64=1.0,
    timestep_name::Symbol=:Δt,
    free_time::Bool=false,
    timestep::Union{Float64,Symbol}=free_time ? timestep_name : timestep_value,
    state_dim::Int=3,
    control_dim::Int=2
)

Create a random NamedTrajectory with T time steps, a state variable x of dimension state_dim, and a control variable u of dimension control_dim. If free_time is true, the time step is a symbol timestep_name and the time step value is timestep_value. Otherwise, the time step is a number timestep_value.